requirements.txt
streamlit
openai
python-dotenv
pandas>=2.0.0
beautifulsoup4>=4.12.0
pylatexenc>=2.10
pdfminer.six>=20221105
lxml>=4.9.0
python-magic>=0.4.27
langchain
langchain_community
langchain_openai
